renamed addItem to addInventory
